+++ Runtime/Mono/ContainerNode.cs
@@ -1,4 +1,5 @@
-﻿using UnityEngine;
+﻿using System.Linq;
+using UnityEngine;
 
 namespace AceLand.NodeFramework.Mono
 {
@@ -8,5 +9,18 @@ namespace AceLand.NodeFramework.Mono
         // empty MonoNode
         // use on structure-use function-less GameObject
         // for filling up gaps in GameObject structure
+
+        protected override void StartAfterNodeBuilt()
+        {
+            base.StartAfterNodeBuilt();
+
+            if (this.IsRoot())
+            {
+                Debug.Log($"ContainerNode {Id} | Active: {Go.activeInHierarchy} | isRoot | {this.Children().Count()}");
+                return;
+            }
+            
+            Debug.Log($"ContainerNode {Id} | Active: {Go.activeInHierarchy} | {this.Parent().Id} | {this.Children().Count()}");
+        }
     }
 }